refactor: integrate flow operator and snapshot
